# contracts/openapi.json
{
  "components": {
    "schemas": {
...
      "ApiResponseFeedbackAdminView": {
        "properties": {
          "data": {
            "$ref": "#/components/schemas/FeedbackAdminView"
          },
          "message": {
            "type": "string"
          },
          "success": {
            "type": "boolean"
          }
        },
        "required": [
          "data",
          "message",
          "success"
        ],
        "type": "object"
      },
...
      "ApiResponseListFeedbackAssignee": {
        "properties": {
          "data": {
            "items": {
              "$ref": "#/components/schemas/FeedbackAssignee"
            },
            "type": "array"
          },
          "message": {
            "type": "string"
          },
          "success": {
            "type": "boolean"
          }
        },
        "required": [
          "data",
          "message",
          "success"
        ],
        "type": "object"
      },
...
      "ApiResponsePageResultFeedbackAdminView": {
        "properties": {
          "data": {
            "$ref": "#/components/schemas/PageResultFeedbackAdminView"
          },
          "message": {
            "type": "string"
          },
          "success": {
            "type": "boolean"
          }
        },
        "required": [
          "data",
          "message",
          "success"
        ],
        "type": "object"
      },
...
      "FeedbackAdminView": {
        "properties": {
          "articleId": {
            "format": "int64",
            "type": "integer"
          },
          "assigneeId": {
            "format": "int64",
            "type": "integer"
          },
          "assigneeName": {
            "type": "string"
          },
          "contact": {
            "type": "string"
          },
          "content": {
            "type": "string"
          },
          "createdAt": {
            "format": "date-time",
            "type": "string"
          },
          "history": {
            "items": {
              "$ref": "#/components/schemas/FeedbackHistory"
            },
            "type": "array"
          },
          "id": {
            "format": "int64",
            "type": "integer"
          },
          "status": {
            "type": "string"
          },
          "title": {
            "type": "string"
          },
          "type": {
            "type": "string"
          },
          "version": {
            "format": "int64",
            "type": "integer"
          }
        },
        "required": [
          "content",
          "createdAt",
          "history",
          "id",
          "status",
          "title",
          "type",
          "version"
        ],
        "type": "object"
      },
      "FeedbackAssignee": {
        "properties": {
          "label": {
            "type": "string"
          },
          "value": {
            "format": "int64",
            "type": "integer"
          }
        },
        "required": [
          "label",
          "value"
        ],
        "type": "object"
      },
      "FeedbackHistory": {
        "properties": {
          "actor": {
            "type": "string"
          },
          "createdAt": {
            "format": "date-time",
            "type": "string"
          },
          "feedbackId": {
            "format": "int64",
            "type": "integer"
          },
          "id": {
            "format": "int64",
            "type": "integer"
          },
          "internalNote": {
            "type": "string"
          },
          "publicReply": {
            "type": "string"
          },
          "status": {
            "type": "string"
          },
          "updatedAt": {
            "format": "date-time",
            "type": "string"
          },
          "version": {
            "format": "int64",
            "type": "integer"
          }
        },
        "required": [
          "actor",
          "createdAt",
          "feedbackId",
          "id",
          "status",
          "version"
        ],
        "type": "object"
      },
      "FeedbackProcess": {
        "properties": {
          "assigneeId": {
            "format": "int64",
            "type": "integer"
          },
          "internalNote": {
            "maxLength": 2000,
            "minLength": 0,
            "type": "string"
          },
          "publicReply": {
            "maxLength": 2000,
            "minLength": 0,
            "type": "string"
          },
          "status": {
            "minLength": 1,
            "pattern": "OPEN|PROCESSING|RESOLVED|CLOSED",
            "type": "string"
          },
          "version": {
            "format": "int64",
            "type": "integer"
          }
        },
        "required": [
          "status",
          "version"
        ],
        "type": "object"
      },
...
      "PageResultFeedbackAdminView": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/FeedbackAdminView"
            },
            "type": "array"
          },
          "page": {
            "format": "int32",
            "type": "integer"
          },
          "size": {
            "format": "int32",
            "type": "integer"
          },
          "total": {
            "format": "int64",
            "type": "integer"
          }
        },
        "type": "object"
      },
...
    },
    "securitySchemes": {
      "bearerAuth": {
        "description": "通过滑块验证与登录接口取得的一次性会话令牌",
        "scheme": "bearer",
        "type": "http"
      }
    }
  },
  "info": {
    "description": "统一响应、动作权限、数据范围与乐观锁均由服务端执行。关闭模块不开放接口。",
    "title": "Mayday API",
    "version": "1.0.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/operations/feedback": {
      "get": {
        "operationId": "FeedbackController_list",
        "parameters": [
          {
            "in": "query",
            "name": "keyword",
            "required": false,
            "schema": {
              "default": "",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "status",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "page",
            "required": false,
            "schema": {
              "default": 1,
              "format": "int32",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "size",
            "required": false,
            "schema": {
              "default": 10,
              "format": "int32",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponsePageResultFeedbackAdminView"
                }
              }
            },
            "description": "OK"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "tags": [
          "feedback-controller"
        ]
      }
    },
    "/api/operations/feedback/assignees": {
      "get": {
        "operationId": "FeedbackController_assignees",
        "parameters": [
          {
            "in": "query",
            "name": "keyword",
            "required": false,
            "schema": {
              "default": "",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponseListFeedbackAssignee"
                }
              }
            },
            "description": "OK"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "tags": [
          "feedback-controller"
        ]
      }
    },
    "/api/operations/feedback/{id}": {
      "get": {
        "operationId": "FeedbackController_detail",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "format": "int64",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponseFeedbackAdminView"
                }
              }
            },
            "description": "OK"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "tags": [
          "feedback-controller"
        ]
      }
    },
    "/api/operations/feedback/{id}/process": {
      "post": {
        "operationId": "FeedbackController_process",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "format": "int64",
              "type": "integer"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/FeedbackProcess"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponseFeedbackAdminView"
                }
              }
            },
            "description": "OK"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "tags": [
          "feedback-controller"
        ]
      }
    },
...
  },
  "servers": [
    {
      "url": "/"
    }
  ]
}